requirements.txt
pydantic>=2.5.0
python-dotenv>=1.0.0
streamlit>=1.31.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-qdrant>=0.0.1
langchain-huggingface>=0.0.1
langchain-ollama>=0.0.1
langchain-text-splitters>=0.0.1
sentence-transformers>=2.2.2
qdrant-client>=1.7.0
pypdf>=3.17.1
python-docx>=1.0.1
python-pptx>=0.6.21
openpyxl>=3.1.2
beautifulsoup4>=4.12.0
markdown>=3.5.0
chardet>=5.2.0
